Cleaned up property estimator test file. Added examples of starting a client, server, and both at the same time. Removed the rdkit docs dependency.
