+++ docs/environment.yml
@@ -2,7 +2,6 @@ name: openforcefield
 channels:
     - omnia
     - conda-forge
-    - rdkit
 dependencies:
     - python ==3.6
     - numpy
@@ -19,7 +18,6 @@ dependencies:
     - nbsphinx
     - sphinx_bootstrap_theme
     - m2r >=0.2.1
-    - rdkit
     - testpath =0.3.1
     - packmol
     - pymbar
