+++ examples/property_estimator/property-data/e2714683-1543-43a3-9406-0b521fca5c63|build_coordinates/e2714683-1543-43a3-9406-0b521fca5c63|build_topology/e2714683-1543-43a3-9406-0b521fca5c63|energy_minimisation/e2714683-1543-43a3-9406-0b521fca5c63|npt_equilibration/e2714683-1543-43a3-9406-0b521fca5c63|converge_uncertainty/e2714683-1543-43a3-9406-0b521fca5c63|npt_production/statistics.dat
@@ -1,0 +1,11 @@
+#"Step","Potential Energy (kJ/mole)","Temperature (K)","Box Volume (nm^3)"
+20,6238.890238554784,188.20681180041845,26.511575131432007
+40,5934.140238554784,202.50821290007804,26.511575131432007
+60,6298.114847929784,191.18203024990945,26.511575131432007
+80,6179.612894804784,200.21248086316882,26.511575131432007
+100,6459.614847929784,189.4669331791827,26.511575131432007
+120,6101.257426054784,208.93348241085258,26.511575131432007
+140,6160.108988554784,211.6834255778983,26.511575131432007
+160,6606.823832304784,193.8665412393603,26.511575131432007
+180,6652.437113554784,197.7487653692238,26.511575131432007
+200,6168.546488554784,223.78398591004336,26.511575131432007
