+++ examples/property_estimator/property-data/e2714683-1543-43a3-9406-0b521fca5c63|build_coordinates/e2714683-1543-43a3-9406-0b521fca5c63|build_topology/e2714683-1543-43a3-9406-0b521fca5c63|energy_minimisation/e2714683-1543-43a3-9406-0b521fca5c63|npt_equilibration/statistics.dat
@@ -1,0 +1,3 @@
+#"Step","Potential Energy (kJ/mole)","Temperature (K)","Box Volume (nm^3)"
+1,2648.2222698047844,264.8077779502668,26.511575131432007
+2,3774.1793010547844,213.2369577933993,26.511575131432007
